v0.1.0 - Nested & Getters finished, begin testing
Nested & Getters finished, begin testing

+++ Cardgame Framework/Assets/CGEngine/Scripts/Classes/CardField.cs
@@ -8,7 +8,7 @@ namespace CardGameFramework
 	{
 		public string fieldName;
 		public CardFieldDataType dataType;
-		public double numValue;
+		public float numValue;
 		public string stringValue;
 		public Sprite imageValue;
 		public string imageSourceName;
